+++ src/main/java/com/kdt/instakyuram/util/ImageManager.java
@@ -15,8 +15,6 @@ import org.slf4j.LoggerFactory;
 import org.springframework.core.io.FileSystemResource;
 import org.springframework.web.multipart.MultipartFile;
 
-import com.kdt.instakyuram.exception.InvalidFileException;
-import com.kdt.instakyuram.exception.NotFoundException;
 import com.kdt.instakyuram.post.domain.PostImage;
 
 public class ImageManager {
